+++ HISTORY.md
@@ -6,6 +6,9 @@
 - Setup automated test-runners
     - mkdocs (#7)
     - pytest (#5)
+- Added Logger to DPArray class (#4)
+- Create ``Logger`` class (#4)
+- Setup automated test-runners (#5)
 - Remove int support (#4)
 - Create ``DPArray`` class (#3)
 - Initial project setup (#1, #2)